+++ include/libgpualg/mult.cuh
@@ -1,6 +1,6 @@
 #pragma once
 
-void matrixMultiplication(double* A, double* B, double* C,
+void matrixMultiplication(float* A, float* B, float* C,
     int numARows, int numAColumns,
     int numBRows, int numBColumns,
     int numCRows, int numCColumns);
